pyproject.toml
[project]
name = "gpt-agent-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "azure-identity>=1.24.0",
    "fastapi>=0.116.1",
    "fastmcp>=2.11.3",
    "google-generativeai>=0.8.5",
    "mcp[cli]>=1.13.1",
    "python-jose>=3.5.0",
    "starlette-context>=0.4.0",
]
